requirements.txt
pyflink
kafka-python
psycopg2-binary
numpy
protobuf
streamlit
scikit-learn
joblib
xgboost
matplotlib